# openapi.yaml
...
openapi: "3.1.0"
...
info:
  title: "NomiFun Model Gateway Protocol v1"
  version: "1.0"
  summary: "Independent community-operated gateway contract and M0 synthetic mock"
  description: "NomiFun official does not operate gateway instances, sell API access or run paid services. Community partners independently operate instances. Clients configure the base URL at runtime. This document freezes control-plane fields and native passthrough boundaries; opaque native schemas are not exhaustive upstream validators. Unknown additive control fields are accepted. credentials cannot enter URLs except the sole inbound Gemini ?key compatibility on its two native routes; never use query credentials in desktop clients, examples, redirects or logs. Auth conflicts/malformed sources fail401. All responses require x-request-id and Cache-Control:no-store;429 also Retry-After. See docs/protocol/v1.md for semantic and affinity obligations."
  license:
    name: "Apache License 2.0"
    identifier: "Apache-2.0"
servers:
  -
    url: "/"
    description: "Runtime-configured deployment API root; no official operated instance or compiled-in commercial URL."
...
security:
  -
    BearerAuth: []
  -
    XAPIKeyAuth: []
  -
    XGoogAPIKeyAuth: []
paths:
  "/nomifun/v1/meta":
    get:
      operationId: "getMeta"
      tags:
        - "Control plane"
      summary: "Read anonymous instance metadata"
      security: []
      responses:
        "200":
          description: "Success. Required fields retain exact v1 semantics and unknown additive fields are accepted."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Meta"
        "400":
          $ref: "#/components/responses/ControlFailure"
        "401":
          $ref: "#/components/responses/ControlFailure"
        "402":
          $ref: "#/components/responses/ControlFailure"
        "403":
          $ref: "#/components/responses/ControlFailure"
        "404":
          $ref: "#/components/responses/ControlFailure"
        "405":
          $ref: "#/components/responses/ControlFailure"
        "413":
          $ref: "#/components/responses/ControlFailure"
        "415":
          $ref: "#/components/responses/ControlFailure"
        "429":
          $ref: "#/components/responses/ControlRateLimited"
        "500":
          $ref: "#/components/responses/ControlFailure"
        "501":
          $ref: "#/components/responses/ControlFailure"
  "/nomifun/v1/catalog":
    get:
      operationId: "getCatalog"
      tags:
        - "Control plane"
      summary: "Read models authorized for current key"
      responses:
        "200":
          description: "Success. Required fields retain exact v1 semantics and unknown additive fields are accepted."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Catalog"
        "400":
          $ref: "#/components/responses/ControlFailure"
        "401":
          $ref: "#/components/responses/ControlFailure"
        "402":
          $ref: "#/components/responses/ControlFailure"
        "403":
          $ref: "#/components/responses/ControlFailure"
        "404":
          $ref: "#/components/responses/ControlFailure"
        "405":
          $ref: "#/components/responses/ControlFailure"
        "413":
          $ref: "#/components/responses/ControlFailure"
        "415":
          $ref: "#/components/responses/ControlFailure"
        "429":
          $ref: "#/components/responses/ControlRateLimited"
        "500":
          $ref: "#/components/responses/ControlFailure"
        "501":
          $ref: "#/components/responses/ControlFailure"
  "/nomifun/v1/account":
    get:
      operationId: "getAccount"
      tags:
        - "Control plane"
      summary: "Read account snapshot for current key"
      responses:
        "200":
          description: "Success. Required fields retain exact v1 semantics and unknown additive fields are accepted."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/ControlFailure"
        "401":
          $ref: "#/components/responses/ControlFailure"
        "402":
          $ref: "#/components/responses/ControlFailure"
        "403":
          $ref: "#/components/responses/ControlFailure"
        "404":
          $ref: "#/components/responses/ControlFailure"
        "405":
          $ref: "#/components/responses/ControlFailure"
        "413":
          $ref: "#/components/responses/ControlFailure"
        "415":
          $ref: "#/components/responses/ControlFailure"
        "429":
          $ref: "#/components/responses/ControlRateLimited"
        "500":
          $ref: "#/components/responses/ControlFailure"
        "501":
          $ref: "#/components/responses/ControlFailure"
  "/v1/models":
    get:
      operationId: "listNativeModels"
      tags:
        - "Native inference"
      summary: "OpenAI-compatible visible model list; IDs match authorized catalog"
      responses:
        "200":
          description: "Native success: opaque upstream JSON; preserve unknown fields. Streaming operations preserve raw native SSE data and event ordering."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/NativeBody"
        "401":
          $ref: "#/components/responses/OpenAIFailure"
        "402":
          $ref: "#/components/responses/OpenAIFailure"
        "403":
          $ref: "#/components/responses/OpenAIFailure"
        "404":
          $ref: "#/components/responses/OpenAIFailure"
        "405":
          $ref: "#/components/responses/OpenAIFailure"
        "429":
          $ref: "#/components/responses/OpenAIRateLimited"
        default:
          $ref: "#/components/responses/OpenAIFailure"
  "/v1/chat/completions":
    post:
      operationId: "createChatCompletion"
      tags:
        - "Native inference"
      summary: "Native openai passthrough"
      description: "Native payload validation remains native. No cross-protocol conversion. Unknown fields and stream deltas/signatures must survive production relay."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/NativeBody"
      responses:
        "200":
          description: "Native success: opaque upstream JSON; preserve unknown fields. Streaming operations preserve raw native SSE data and event ordering."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/NativeBody"
            "text/event-stream":
              schema:
                $ref: "#/components/schemas/RawSSE"
              x-sse-event-schema:
                $ref: "#/components/schemas/ChatChunk"
              x-sse-terminal-value: "[DONE]"
        "400":
          $ref: "#/components/responses/OpenAIFailure"
        "401":
          $ref: "#/components/responses/OpenAIFailure"
        "402":
          $ref: "#/components/responses/OpenAIFailure"
        "403":
          $ref: "#/components/responses/OpenAIFailure"
        "404":
          $ref: "#/components/responses/OpenAIFailure"
        "405":
          $ref: "#/components/responses/OpenAIFailure"
        "413":
          $ref: "#/components/responses/OpenAIFailure"
        "415":
          $ref: "#/components/responses/OpenAIFailure"
        "429":
          $ref: "#/components/responses/OpenAIRateLimited"
        "500":
          $ref: "#/components/responses/OpenAIFailure"
        "501":
          $ref: "#/components/responses/OpenAIFailure"
        "502":
          $ref: "#/components/responses/OpenAIFailure"
        "503":
          $ref: "#/components/responses/OpenAIFailure"
        "504":
          $ref: "#/components/responses/OpenAIFailure"
        default:
          $ref: "#/components/responses/OpenAIFailure"
  "/v1/responses":
    post:
      operationId: "createResponse"
      tags:
        - "Native inference"
      summary: "Native openai-response passthrough"
      description: "Preserve store/previous_response_id/include reasoning.encrypted_content. Key+response-ID bound to the original upstream; unknown/cross-key/lost affinity fails HTTP400 invalid_previous_response_id without account failover. Native payload validation remains native. No cross-protocol conversion. Unknown fields and stream deltas/signatures must survive production relay."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/NativeBody"
      responses:
        "200":
          description: "Native success: opaque upstream JSON; preserve unknown fields. Streaming operations preserve raw native SSE data and event ordering."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/NativeBody"
            "text/event-stream":
              schema:
                $ref: "#/components/schemas/RawSSE"
              x-sse-event-schema:
                $ref: "#/components/schemas/ResponseEvent"
        "400":
          $ref: "#/components/responses/OpenAIFailure"
        "401":
          $ref: "#/components/responses/OpenAIFailure"
        "402":
          $ref: "#/components/responses/OpenAIFailure"
        "403":
          $ref: "#/components/responses/OpenAIFailure"
        "404":
          $ref: "#/components/responses/OpenAIFailure"
        "405":
          $ref: "#/components/responses/OpenAIFailure"
        "413":
          $ref: "#/components/responses/OpenAIFailure"
        "415":
          $ref: "#/components/responses/OpenAIFailure"
        "429":
          $ref: "#/components/responses/OpenAIRateLimited"
        "500":
          $ref: "#/components/responses/OpenAIFailure"
        "501":
          $ref: "#/components/responses/OpenAIFailure"
        "502":
          $ref: "#/components/responses/OpenAIFailure"
        "503":
          $ref: "#/components/responses/OpenAIFailure"
        "504":
          $ref: "#/components/responses/OpenAIFailure"
        default:
          $ref: "#/components/responses/OpenAIFailure"
  "/v1/embeddings":
    post:
      operationId: "createEmbeddings"
      tags:
        - "Native inference"
      summary: "Native embeddings passthrough"
      description: "Native payload validation remains native. No cross-protocol conversion. Unknown fields and stream deltas/signatures must survive production relay."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/NativeBody"
      responses:
        "200":
          description: "Native success: opaque upstream JSON; preserve unknown fields. Streaming operations preserve raw native SSE data and event ordering."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/NativeBody"
        "400":
          $ref: "#/components/responses/OpenAIFailure"
        "401":
          $ref: "#/components/responses/OpenAIFailure"
        "402":
          $ref: "#/components/responses/OpenAIFailure"
        "403":
          $ref: "#/components/responses/OpenAIFailure"
        "404":
          $ref: "#/components/responses/OpenAIFailure"
        "405":
          $ref: "#/components/responses/OpenAIFailure"
        "413":
          $ref: "#/components/responses/OpenAIFailure"
        "415":
          $ref: "#/components/responses/OpenAIFailure"
        "429":
          $ref: "#/components/responses/OpenAIRateLimited"
        "500":
          $ref: "#/components/responses/OpenAIFailure"
        "501":
          $ref: "#/components/responses/OpenAIFailure"
        "502":
          $ref: "#/components/responses/OpenAIFailure"
        "503":
          $ref: "#/components/responses/OpenAIFailure"
        "504":
          $ref: "#/components/responses/OpenAIFailure"
        default:
          $ref: "#/components/responses/OpenAIFailure"
  "/v1/images/generations":
    post:
      operationId: "generateImage"
      tags:
        - "Native inference"
      summary: "Native image-generation passthrough"
      description: "Native payload validation remains native. No cross-protocol conversion. Unknown fields and stream deltas/signatures must survive production relay."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/NativeBody"
      responses:
        "200":
          description: "Native success: opaque upstream JSON; preserve unknown fields. Streaming operations preserve raw native SSE data and event ordering."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/NativeBody"
        "400":
          $ref: "#/components/responses/OpenAIFailure"
        "401":
          $ref: "#/components/responses/OpenAIFailure"
        "402":
          $ref: "#/components/responses/OpenAIFailure"
        "403":
          $ref: "#/components/responses/OpenAIFailure"
        "404":
          $ref: "#/components/responses/OpenAIFailure"
        "405":
          $ref: "#/components/responses/OpenAIFailure"
        "413":
          $ref: "#/components/responses/OpenAIFailure"
        "415":
          $ref: "#/components/responses/OpenAIFailure"
        "429":
          $ref: "#/components/responses/OpenAIRateLimited"
        "500":
          $ref: "#/components/responses/OpenAIFailure"
        "501":
          $ref: "#/components/responses/OpenAIFailure"
        "502":
          $ref: "#/components/responses/OpenAIFailure"
        "503":
          $ref: "#/components/responses/OpenAIFailure"
        "504":
          $ref: "#/components/responses/OpenAIFailure"
        default:
          $ref: "#/components/responses/OpenAIFailure"
  "/v1/images/edits":
    post:
      operationId: "editImage"
      tags:
        - "Native inference"
      summary: "Native image-generation passthrough"
      description: "Native payload validation remains native. No cross-protocol conversion. Unknown fields and stream deltas/signatures must survive production relay."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/NativeBody"
          "multipart/form-data":
            schema:
              type: "object"
              additionalProperties: true
              description: "Native multipart form including binary images; this is intentionally not an exhaustive image API schema."
      responses:
        "200":
          description: "Native success: opaque upstream JSON; preserve unknown fields. Streaming operations preserve raw native SSE data and event ordering."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/NativeBody"
        "400":
          $ref: "#/components/responses/OpenAIFailure"
        "401":
          $ref: "#/components/responses/OpenAIFailure"
        "402":
          $ref: "#/components/responses/OpenAIFailure"
        "403":
          $ref: "#/components/responses/OpenAIFailure"
        "404":
          $ref: "#/components/responses/OpenAIFailure"
        "405":
          $ref: "#/components/responses/OpenAIFailure"
        "413":
          $ref: "#/components/responses/OpenAIFailure"
        "415":
          $ref: "#/components/responses/OpenAIFailure"
        "429":
          $ref: "#/components/responses/OpenAIRateLimited"
        "500":
          $ref: "#/components/responses/OpenAIFailure"
        "501":
          $ref: "#/components/responses/OpenAIFailure"
        "502":
          $ref: "#/components/responses/OpenAIFailure"
        "503":
          $ref: "#/components/responses/OpenAIFailure"
        "504":
          $ref: "#/components/responses/OpenAIFailure"
        default:
          $ref: "#/components/responses/OpenAIFailure"
  "/v1/rerank":
    post:
      operationId: "rerank"
      tags:
        - "Native inference"
      summary: "Native jina-rerank passthrough"
      description: "Native payload validation remains native. No cross-protocol conversion. Unknown fields and stream deltas/signatures must survive production relay."
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/NativeBody"
      responses:
        "200":
          description: "Native success: opaque upstream JSON; preserve unknown fields. Streaming operations preserve raw native SSE data and event ordering."
          headers:
            x-request-id:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/NativeBody"
        "400":
          $ref: "#/components/responses/OpenAIFailure"
        "401":
          $ref: "#/components/responses/OpenAIFailure"
        "402":
          $ref: "#/components/responses/OpenAIFailure"
        "403":
          $ref: "#/components/responses/OpenAIFailure"
        "404":
          $ref: "#/components/responses/OpenAIFailure"
        "405":
          $ref: "#/components/responses/OpenAIFailure"
        "413":
          $ref: "#/components/responses/OpenAIFailure"
        "415":
          $ref: "#/components/responses/OpenAIFailure"
        "429":
          $ref: "#/components/responses/OpenAIRateLimited"
        "500":
          $ref: "#/components/responses/OpenAIFailure"
        "501":
          $ref: "#/components/responses/OpenAIFailure"
        "502":
          $ref: "#/components/responses/OpenAIFailure"
        "503":
          $ref: "#/components/responses/OpenAIFailure"
        "504":
          $ref: "#/components/responses/OpenAIFailure"
        default:
          $ref: "#/components/responses/OpenAIFailure"
...
components:
  securitySchemes:
    BearerAuth:
      type: "http"
      scheme: "bearer"
      description: "Instance API key; same key accepted by each header alternative. Conflicting/malformed/repeated sources reject401."
    XAPIKeyAuth:
      type: "apiKey"
      in: "header"
      name: "x-api-key"
      description: "Same instance API key accepted on every protected endpoint."
    XGoogAPIKeyAuth:
      type: "apiKey"
      in: "header"
      name: "x-goog-api-key"
      description: "Same instance API key accepted on every protected endpoint."
    GeminiQueryKeyAuth:
      type: "apiKey"
      in: "query"
      name: "key"
      description: "Sole legacy inbound URL-key exception on the two native Gemini routes. Use headers for desktop/examples. Query key must never appear in access logs or redirects; other routes reject key."
  headers:
    RequestID:
      required: true
      description: "Non-empty credential-free request identifier."
      schema:
        type: "string"
        minLength: 1
    NoStore:
      required: true
      description: "All responses must forbid caching of account/model/error data."
      schema:
        type: "string"
        const: "no-store"
...
  responses:
...
    ControlFailure:
      description: "Gateway-generated control-plane error."
      headers:
        x-request-id:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/OpenAIGatewayError"
    ControlRateLimited:
      description: "rate_limited; honor Retry-After."
      headers:
        x-request-id:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStore"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/OpenAIGatewayError"
  schemas:
...
    Meta:
      type: "object"
      required:
        - "contract_version"
        - "operator"
        - "capabilities"
        - "optional_endpoints"
      properties:
        contract_version:
          const: "1.0"
          type: "string"
        operator:
          $ref: "#/components/schemas/Operator"
        capabilities:
          type: "array"
          items:
            $ref: "#/components/schemas/Endpoint"
          uniqueItems: true
        optional_endpoints:
          type: "array"
          items:
            type: "string"
            enum:
              - "/v1/messages/count_tokens"
          uniqueItems: true
      additionalProperties: true
...
    RawSSE:
      type: "string"
      description: "Raw native SSE stream; do not normalize event names, data or terminal markers."
...